requirements.txt
websockets==11.0.3
deepgram-sdk==3.0.3
sounddevice==0.4.6
numpy==1.24.3
vaderSentiment==3.3.2
sentence-transformers==2.2.2
transformers==4.30.2
torch==2.0.1
scikit-learn==1.3.0
asyncio 